定义源代码生成静态类
