+++ src/TelescopeSourceGenerator/Analyzers/ExportTypeToMethod_/IExportMethodCodeGenerator.cs
@@ -4,6 +4,9 @@ using Microsoft.CodeAnalysis;
 
 namespace dotnetCampus.Telescope.SourceGeneratorAnalyzers;
 
+/// <summary>
+/// 导入方法的生成器
+/// </summary>
 interface IExportMethodCodeGenerator
 {
     string GenerateSourceCode(ExportMethodReturnTypeCollectionResult exportMethodReturnTypeCollectionResult,
